+++ packages/common/src/schemas/documents/merchant-screening-plugin-schema.ts
@@ -113,11 +113,13 @@ const PrincipalSchema = Type.Object({
     maxLength: 40,
   }),
   Address: AddressSchema,
-  PhoneNumber: Type.String({
-    description: "The principal owner's phone number, including the area code.",
-    example: '3165557625',
-    maxLength: 25,
-  }),
+  PhoneNumber: Type.Optional(
+    Type.String({
+      description: "The principal owner's phone number, including the area code.",
+      example: '3165557625',
+      maxLength: 25,
+    }),
+  ),
   AltPhoneNumber: Type.Optional(
     Type.String({
       description: "The principal owner's alternate phone number, including the area code.",
@@ -213,11 +215,13 @@ const MerchantSchema = Type.Object({
     }),
   ),
   Address: Type.Optional(AddressSchema),
-  PhoneNumber: Type.String({
-    description: "The Business or Merchant's phone number, including the area code.",
-    example: '3165557625',
-    maxLength: 25,
-  }),
+  PhoneNumber: Type.Optional(
+    Type.String({
+      description: "The Business or Merchant's phone number, including the area code.",
+      example: '3165557625',
+      maxLength: 25,
+    }),
+  ),
   AltPhoneNumber: Type.Optional(
     Type.String({
       description: "The Business or Merchant's alternate phone number, including the area code.",