+++ services/workflows-service/src/workflow/schemas/zod-schemas.ts
@@ -72,6 +72,15 @@ export const ConfigSchema = z
           .optional(),
       })
       .optional(),
+    editableContext: z
+      .object({
+        kyc: z
+          .object({
+            entity: z.boolean().optional(),
+          })
+          .optional(),
+      })
+      .optional(),
   })
   .strict()
   .optional();